requirements.txt

PyYAML
pytest

cloudshell-logging==1.0.3
cloudshell-orch-core>=3.3.0.0,<3.4.0.0
cloudshell-shell-core>=5.0.3,<6.0.0
cloudshell-rest-api>=8.2.3.1
shellfoundry
